Re-organize routes so that split into separate endpoint files.
