+++ package.json
@@ -2,7 +2,7 @@
   "name": "sinopia_api",
   "version": "0.1.0",
   "description": "API for persisting resources in the Sinopia framework.",
-  "main": "server.js",
+  "main": "src/server.js",
   "license": "Apache-2.0",
   "homepage": "https://github.com/LD4P/sinopia_api",
   "repository": "github:LD4P/sinopia_api",
@@ -28,6 +28,8 @@
     "rdf-ext": "^1.3.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.11.4",
+    "@babel/node": "^7.10.5",
     "@babel/preset-env": "^7.11.0",
     "babel-jest": "^26.3.0",
     "eslint": "^7.7.0",
@@ -43,9 +45,6 @@
     "moduleDirectories": [
       "node_modules",
       "src"
-    ],
-    "transform": {
-      "^.+\\.[jt]sx?$": "babel-jest"
-    }
+    ]
   }
 }
